removed redirects from profile
