+++ src/App.js
@@ -47,7 +47,7 @@ function App() {
             <Route exact path="/login" component={Login} />
             <UserRoute exact path={["/", "/orders"]} component={OrdersList} />
             <UserRoute path="/print/:id" component={PrintOrder} />
-            <UserRoute exact path="/profile" component={Profile} />
+            <Route exact path="/profile" component={Profile} />
             <Route path="/scan/:id" component={ScanOrder} />
             <EditorRoute path="/orders/:id" component={EditOrder} />
             <AdminRoute exact path="/users/add" component={AddUser} />
